requirements.txt
# Application Streamlit basique - MNIST
# Version simplifiée

tensorflow
streamlit
Pillow
numpy
tensorflow-datasets

